Slightly refactor otp_brand CLI
